requirements.txt
aiogram
fastapi
uvicorn
python-docx
PyPDF2
requests